refactor: reoder (#3)
