+++ .gitignore
@@ -2,4 +2,4 @@
 .nyc_output
 coverage
 node_modules
-test-results.json
+report
